requirements.txt
google-api-python-client==1.8.0
grpcio==1.27.2
grpcio-tools==1.27.2
PyYAML==5.3